ENH: copy previous items content when forking by default
handle items with subdirs when copying

+++ models.py
@@ -222,15 +222,20 @@ def import_item(
         tmp_dir = os.path.join(repository, kimcode)
         tarfile_obj.extractall(path=tmp_dir)
         contents = os.listdir(tmp_dir)
-        # print(contents)
         # if the contents of the item are enclosed in a directory, copy them out
         # then delete the directory
-        if len(contents) == 1:
+        if len(contents) == 1 and os.path.isdir(os.path.join(tmp_dir, contents[0])):
             inner_dir = os.path.join(tmp_dir, contents[0])
             if os.path.isdir(inner_dir):
                 inner_contents = os.listdir(inner_dir)
                 for item in inner_contents:
-                    shutil.copy(os.path.join(inner_dir, item), tmp_dir)
+                    item_path = os.path.join(inner_dir, item)
+                    if os.path.isdir(item_path):
+                        shutil.copytree(
+                            item_path, os.path.join(tmp_dir, item), dirs_exist_ok=True
+                        )
+                    else:
+                        shutil.copy(os.path.join(inner_dir, item), tmp_dir)
                 shutil.rmtree(inner_dir)
 
         executables = []
@@ -437,7 +442,7 @@ def version_update(
     """
 
     this_user = users.whoami()
-    if users.is_user(system_username=this_user):
+    if users.is_user(username=this_user):
         UUID = users.get_user_info(username=this_user)["uuid"]
     else:
         raise cf.KIMkitUserNotFoundError(
@@ -499,12 +504,18 @@ def version_update(
         contents = os.listdir(tmp_dir)
         # if the contents of the item are enclosed in a directory, copy them out
         # then delete the directory
-        if len(contents) == 1:
+        if len(contents) == 1 and os.path.isdir(os.path.join(tmp_dir, contents[0])):
             inner_dir = os.path.join(tmp_dir, contents[0])
             if os.path.isdir(inner_dir):
                 inner_contents = os.listdir(inner_dir)
                 for item in inner_contents:
-                    shutil.copy(os.path.join(inner_dir, item), tmp_dir)
+                    item_path = os.path.join(inner_dir, item)
+                    if os.path.isdir(item_path):
+                        shutil.copytree(
+                            item_path, os.path.join(tmp_dir, item), dirs_exist_ok=True
+                        )
+                    else:
+                        shutil.copy(os.path.join(inner_dir, item), tmp_dir)
                 shutil.rmtree(inner_dir)
 
         executables = []
@@ -570,7 +581,7 @@ def version_update(
 def fork(
     kimcode,
     new_kimcode,
-    tarfile_obj,
+    tarfile_obj=None,
     repository=cf.LOCAL_REPOSITORY_PATH,
     metadata_update_dict=None,
     provenance_comments=None,
@@ -586,7 +597,7 @@ def fork(
         ID code of the item to be forked
     new_kimcode : str
         id code the new item will be assigned
-    tarfile_obj : tarfile.Tarfile
+    tarfile_obj : tarfile.Tarfile, optional
         tarfile object containing the new version's content
     repository : path-like, optional
         root directory of the KIMkit repo containing the item,
@@ -611,7 +622,7 @@ def fork(
     """
 
     this_user = users.whoami()
-    if users.is_user(system_username=this_user):
+    if users.is_user(username=this_user):
         UUID = users.get_user_info(username=this_user)["uuid"]
     else:
         raise cf.KIMkitUserNotFoundError(
@@ -652,16 +663,30 @@ def fork(
         kim_item_type = "model-driver"
 
     tmp_dir = os.path.join(repository, new_kimcode)
-    tarfile_obj.extractall(path=tmp_dir)
+    if tarfile_obj:
+        tarfile_obj.extract(path=tmp_dir)
+    else:
+        # copy the existing item without editing it
+        # if no new content supplied
+        tarfile_obj = export(kimcode)
+        for item in tarfile_obj:
+            if kimcode in item.getnames():
+                item.extractall(path=tmp_dir)
     contents = os.listdir(tmp_dir)
     # if the contents of the item are enclosed in a directory, copy them out
     # then delete the directory
-    if len(contents) == 1:
+    if len(contents) == 1 and os.path.isdir(os.path.join(tmp_dir, contents[0])):
         inner_dir = os.path.join(tmp_dir, contents[0])
         if os.path.isdir(inner_dir):
             inner_contents = os.listdir(inner_dir)
             for item in inner_contents:
-                shutil.copy(os.path.join(inner_dir, item), tmp_dir)
+                item_path = os.path.join(inner_dir, item)
+                if os.path.isdir(item_path):
+                    shutil.copytree(
+                        item_path, os.path.join(tmp_dir, item), dirs_exist_ok=True
+                    )
+                else:
+                    shutil.copy(os.path.join(inner_dir, item), tmp_dir)
             shutil.rmtree(inner_dir)
 
     executables = []
